Remove SkillService interface, add Skill @Entity annotation
